Set up index and style pages
I don't quite know what to say. I tried my best to
complete the task required. I'm don't think it is
the perfect setup, but it looked okay on my pc.

I probably need to redo the sizing and fix the
footer. Other than that, for the most part, the
bulk of the work has been already done.

Really looking forward to JS part of the foundation
course!